Update certificate dirs and file names on FreeBSD
FreeBSD contains a canonical certstore managed by certctl(8) since 12.2 located
in the base system (/etc/ssl), search there first. Alternatively, a user can
populate a custom store in distbase (/usr/local/etc/ssl) with certctl(8) which
shall be queried if the former does not exist. Within these there can be also a
bundle in parallel to a hashed directory.

This fixes #20 and fixes #37